+++ torch_geometric/nn/conv/pna_conv.py
@@ -73,11 +73,11 @@ class PNAConv(MessagePassing):
             be split between towers or not (default: :obj:`False`).
         act (str or Callable, optional): Pre- and post-layer activation
             function to use. (default: :obj:`"relu"`)
-        train_norm (bool, optional) Whether a normalization parameter
-            is fixed or a trainable parameter. (default: :obj:`False`)
         act_kwargs (Dict[str, Any], optional): Arguments passed to the
             respective activation function defined by :obj:`act`.
             (default: :obj:`None`)
+        train_norm (bool, optional) Whether normalization parameters
+            are trainable. (default: :obj:`False`)
         **kwargs (optional): Additional arguments of
             :class:`torch_geometric.nn.conv.MessagePassing`.
 
@@ -101,8 +101,8 @@ class PNAConv(MessagePassing):
         post_layers: int = 1,
         divide_input: bool = False,
         act: Union[str, Callable, None] = "relu",
-        train_norm: bool = False,
         act_kwargs: Optional[Dict[str, Any]] = None,
+        train_norm: bool = False,
         **kwargs,
     ):